Fix more warnings

+++ cpu/i386/isr.c
@@ -180,7 +180,7 @@ void isr_handler(registers_t r) {
       } else if (r.eax==5) {
         r.ebx=(uint32_t)tasking_get_errno_address();
       } else if (r.eax==6) {
-        r.ebx=tasking_get_msg((uint32_t*)r.ebx);
+        r.ebx=(uint32_t)tasking_get_msg((uint32_t*)r.ebx);
       } else if (r.eax==7) {
         tasking_send_msg(r.ebx,(void*)r.ecx,r.edx);
       }
